change testing structure
